Renovacao geral de entidades + inicio BD

+++ src/main/java/com/projark/trab/InterfacesAdaptadoras/Repositorios/RepositorioAeronaves.java
@@ -3,10 +3,7 @@ package com.projark.trab.InterfacesAdaptadoras.Repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projark.trab.Dominio.Aeronave;
-import com.projark.trab.Dominio.IAeronavesRep;
 
-public interface RepositorioAeronaves extends JpaRepository<Aeronave,Integer>{
-
-    
+public interface RepositorioAeronaves extends JpaRepository<Aeronave, Long> {
 
 }
